build: integration test fixes
karma config made global
integration tests run alright
