+++ packages/webrtc-core/package.json
@@ -21,7 +21,7 @@
     "build": "tsc && rollup --config",
     "test": "npx jest",
     "start": "web-dev-server --config web-dev-server.config.js",
-    "test:integration": "karma start --integration",
+    "test:integration": "karma start karma.conf.cjs --integration",
     "test:integration:debug": "karma start --integration --browsers=chrome --single-run=false --debug --auto-watch",
     "test:debug": "karma start --browsers=chrome --single-run=false --debug --auto-watch",
     "release": "semantic-release"
